LayoutTests/imported/w3c/web-platform-tests/html/webappapis/dynamic-markup-insertion/closing-the-input-stream/document-close-with-pending-script-expected.txt



PASS document.close called while a script is pending
